Photo viewer, clock in taskbar, misc. updates

+++ drivers/clipdraw/clipdraw/desktop.hpp
@@ -36,7 +36,7 @@ public:
 
 	uint32_t desktopDecode(int val);
 	uint32_t desktopDecodeLow[256];
-	uint16_t* desktopBuffer;
+	uint32_t* desktopBuffer;
 
 	uint64_t invalidateAllJustOccured;
